requirements.txt
selenium
python-dotenv
customtkinter
